BOL-375 refactor: move AuthRepositoryImpl's package
